// src/types/CulturalFacilitiesType.d.ts
export type CulturalFacilitiesDetailInfoType = {
  contentid: string;
  contenttypeid: string;
  serialnum: string;
  infoname: string;
  infotext: string;
  fldgubun: string;
};

export type CulturalFacilitiesDetailIntroType = {
  contentid: string;
  contenttypeid: string;
  scale: string;
  usefee: string;
  discountinfo: string;
  spendtime: string;
  parkingfee: string;
  infocenterculture: string;
  accomcountculture: string;
  usetimeculture: string;
  restdateculture: string;
  parkingculture: string;
  chkbabycarriageculture: string;
  chkpetculture: string;
  chkcreditcardculture: string;
};